// src/app/modules/compensation-admin/models/salary-basis.ts
/*
 * Copyright (c) 2021 All Rights Reserved.
 *
 * Unauthorized copying of this file, via any medium is strictly prohibited.
 * Proprietary and confidential.
 *
 * Written by:
 *   - Satrya Wiguna <satrya@freshcms.net>
 */

export type SalaryBasis = {
  salary_basis_id?: string;
  is_read_only?: boolean;
  salary_basis_code?: string;
  salary_basis_name?: string;
  element_id?: string;
  element_name?: string;
  input_value_id?: string;
  input_value_name?: string;
  description?: string;
};